Refactor feature flag structure
