+++ src/lib/auth/AuthProviders/OAuthAuthenticator/index.ts
@@ -2,8 +2,6 @@ import axios from 'axios'
 import { createHash } from 'crypto'
 import isWsl from 'is-wsl'
 import jwt from 'jsonwebtoken'
-import open from 'open'
-import opn from 'opn'
 import { join } from 'path'
 import logger from '../../../../api/logger'
 import { VTEXID } from '../../../../api/clients/IOClients/external/VTEXID'
@@ -13,9 +11,7 @@ import { randomCryptoString } from '../../../utils/randomCryptoString'
 import { spawnUnblockingChildProcess } from '../../../utils/spawnUnblockingChildProcess'
 import { AuthProviderBase } from '../AuthProviderBase'
 import { LoginServer } from './LoginServer'
-import { ToolbeltConfig } from '../../../../api/clients/IOClients/apps/ToolbeltConfig'
-import { ErrorReport } from '../../../../api/error/ErrorReport'
-import { ErrorKinds } from '../../../../api/error/ErrorKinds'
+import { switchOpen } from '../../../../modules/featureFlagDecider'
 
 export class OAuthAuthenticator extends AuthProviderBase {
   public static readonly AUTH_TYPE = 'oauth'
@@ -43,22 +39,8 @@ export class OAuthAuthenticator extends AuthProviderBase {
 
       loginServer.setLoginState(loginState)
 
-      try {
-        const url = await this.loginUrl(account, loginState)
-        const configClient = ToolbeltConfig.createClient()
-        const { featureFlags } = await configClient.getGlobalConfig()
-
-        if (featureFlags.FEATURE_FLAG_NEW_OPEN_PACKAGE) {
-          open(url, { wait: false })
-        } else {
-          opn(url, { wait: false })
-        }
-      } catch (err) {
-        ErrorReport.createAndMaybeRegisterOnTelemetry({
-          kind: ErrorKinds.TOOLBELT_CONFIG_FEATURE_FLAG_ERROR,
-          originalError: err,
-        }).logErrorForUser({ coreLogLevelDefault: 'debug' })
-      }
+      const url = await this.loginUrl(account, loginState)
+      switchOpen(url, { wait: false })
 
       if (isWsl) {
         logger.warn(
