+++ src/modules/browse.ts
@@ -1,11 +1,7 @@
-import open from 'open'
-import opn from 'opn'
 import QRCode from 'qrcode-terminal'
-import { ToolbeltConfig } from '../api/clients/IOClients/apps/ToolbeltConfig'
-import { ErrorKinds } from '../api/error/ErrorKinds'
-import { ErrorReport } from '../api/error/ErrorReport'
 import { SessionManager } from '../api/session/SessionManager'
 import { storeUrl } from '../api/storeUrl'
+import { switchOpen } from './featureFlagDecider'
 
 interface BrowseOptions {
   qr: boolean
@@ -19,18 +15,5 @@ export default async (path: string, { qr }: BrowseOptions) => {
     QRCode.generate(uri, { small: true })
     return
   }
-  try {
-    const configClient = ToolbeltConfig.createClient()
-    const { featureFlags } = await configClient.getGlobalConfig()
-    if (featureFlags.FEATURE_FLAG_NEW_OPEN_PACKAGE) {
-      open(uri, { wait: false })
-    } else {
-      opn(uri, { wait: false })
-    }
-  } catch (err) {
-    ErrorReport.createAndMaybeRegisterOnTelemetry({
-      kind: ErrorKinds.TOOLBELT_CONFIG_FEATURE_FLAG_ERROR,
-      originalError: err,
-    }).logErrorForUser({ coreLogLevelDefault: 'debug' })
-  }
+  switchOpen(uri, { wait: false })
 }
