+++ src/nps.ts
@@ -1,13 +1,8 @@
 import enquirer from 'enquirer'
 import moment from 'moment'
-import open from 'open'
-import opn from 'opn'
-import { ToolbeltConfig } from './api/clients/IOClients/apps/ToolbeltConfig'
-
+import { switchOpen } from './modules/featureFlagDecider'
 import { getNextFeedbackDate, saveNextFeedbackDate } from './api/conf'
-import { ErrorReport } from './api/error/ErrorReport'
 import { promptConfirm } from './api/modules/prompts'
-import { ErrorKinds } from './api/error/ErrorKinds'
 
 const NPSFormURL = 'https://forms.gle/CRRHn6P3x9AeaWTQ8'
 
@@ -45,22 +40,7 @@ export async function checkAndOpenNPSLink() {
           .add(3, 'months')
           .toISOString()
       )
-
-      try {
-        const configClient = ToolbeltConfig.createClient()
-        const { featureFlags } = await configClient.getGlobalConfig()
-
-        if (featureFlags.FEATURE_FLAG_NEW_OPEN_PACKAGE) {
-          open(NPSFormURL, { wait: false })
-        } else {
-          opn(NPSFormURL, { wait: false })
-        }
-      } catch (err) {
-        ErrorReport.createAndMaybeRegisterOnTelemetry({
-          kind: ErrorKinds.TOOLBELT_CONFIG_FEATURE_FLAG_ERROR,
-          originalError: err,
-        }).logErrorForUser({ coreLogLevelDefault: 'debug' })
-      }
+      switchOpen(NPSFormURL, { wait: false })
     } else {
       // @ts-ignore
       let { remindChoice } = await enquirer.prompt({
